footbal_interface v0.1.0 FootbalInterface.Formatter

Contains the logic to convert messages into the various protocols supported. Currently supports JSON and Google's Protocol Buffer.

Functions

format(results, format)

format(results :: [Map.t()], format :: String.t()) ::
  {:ok, String.t()}
  | {:error, :invalid_format, String.t()}
  | {:error, :json_encoding_failed, any()}
  | {:error, :protobuff_encoding_failed, any()}

Receives a list of results and converts it to either JSON or protobuf.

Arguments:

results: the list of results from a DB query.

format: the format to convert to.

Returns:

{:ok, result}: Tagged :ok tuple with the result in binary.

{:error, :invalid_format, inv_format}: if the requested format is not supported. inv_format is the invalid format.

{:error, :json_encoding_failed, err}: if converting the data to JSON failed. err is the error raised.

{:error, :protobuff_encoding_failed, err}: if converting the data to protobuff failed. err is the error raised.